requirements.txt
autopep8==1.5.7
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
dnspython==2.1.0
email-validator==1.1.2
Flask==2.0.0
Flask-Login==0.5.0
Flask-Principal==0.4.0
Flask-Security-Too==4.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
passlib==1.7.4
pycodestyle==2.7.0
python-dotenv==0.17.1
requests==2.25.1
SQLAlchemy==1.4.15
toml==0.10.2
urllib3==1.26.4
Werkzeug==2.0.1
WTForms==2.3.3
